feedback/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
Django==4.2.6
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-speech==2.21.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
huggingface-hub==0.17.3
HyperPyYAML==1.2.2
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
proglog==0.1.10
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scipy==1.11.3
sentencepiece==0.1.99
six==1.16.0
speechbrain==0.5.15
sqlparse==0.4.4
sympy==1.12
tenacity==8.2.3
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
